refactor: Separate functions in node.ts and utils.ts files (#7)
* chore: ignore out dir

* chore: separate functions in node.ts and utils.ts files

* fix: unified import target
